refactoring
problems to solve:
to properly report ignore pattern list before application,
ignore patterns can't be hidden in compare-directories.ps1

also, ignore pattern application was horrendously inefficient
(for-file -> for-pattern -> for-sourcefile)
now its a single for-sourcefiles.

as a result, the before-pretty-short backup.ps1 is now a lot heavier.
however, Compare-Directories (now Get-ListDiff) is a lot lighter.

open issues:
MatchResult and PatternMatch are in need of comments now.
And the whole thing has not been debugged yet.